Pass TSM_VTE_INVALID for invalid UCS4 values
xkbcommon returns 0 for invalid ucs4 values. Unfortunately, this is a
valid UCS4 code. That's why TSM requires us to pass TSM_VTE_INVALID
instead. So add a short fixup to convert 0 to TSM_VTE_INVALID.

We rely on xkbcommon to never return 0 for valid codes now. This seems
unfortunate, but the API was broken right from the beginning.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>